+++ database/migrations/2025_04_09_213605_create_vehicle_companies_table.php
@@ -11,15 +11,17 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('vehicle_companies', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('vehicle_id')->constrained('vehicles')->onDelete('no action'); // Relación con la tabla vehicles
-            $table->foreignId('company_id')->constrained('companies')->onDelete('no action'); // Relación con la tabla companies
-            $table->boolean('is_linked')->default(true); // Estado de vinculación (true = vinculado, false = desvinculado)
-            $table->foreignId('linked_by')->nullable()->constrained('users')->onDelete('no action'); // Relación con la tabla users
-            $table->foreignId('unlinked_by')->nullable()->constrained('users')->onDelete('no action'); // Relación con la tabla users
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('vehicle_companies')) {
+            Schema::create('vehicle_companies', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('vehicle_id')->constrained('vehicles')->onDelete('no action'); // Relación con la tabla vehicles
+                $table->foreignId('company_id')->constrained('companies')->onDelete('no action'); // Relación con la tabla companies
+                $table->boolean('is_linked')->default(true); // Estado de vinculación (true = vinculado, false = desvinculado)
+                $table->foreignId('linked_by')->nullable()->constrained('users')->onDelete('no action'); // Relación con la tabla users
+                $table->foreignId('unlinked_by')->nullable()->constrained('users')->onDelete('no action'); // Relación con la tabla users
+                $table->timestamps();
+            });
+        }
     }
 
     /**
